requirements/prod.txt
# It is recommended to use a virtual environment.
# Instructions below:

# Before installing these requirements,
# cd into your project directory and
# create a virtual environment (if you did
# not do so earlier) using the command:
# $ python -m venv .venv
# (if you get an error and using conda,
# you might need to update conda and
# then python:
# $ conda update -n base -c defaults conda
# $ conda update python
# )

# Activate it using, on Windows:
# $ .venv\Scripts\activate.bat
# On maxOX/Unix:
# $ source .venv/bin/activate

# THEN install these requirements using the command:
# $ pip -r requirements.txt
# A guide to venv is available here:
# https://docs.python.org/3/tutorial/venv.html


# Library dependencies
# PLEASE KEEP IN SYNC WITH install_requires in setup.py!
flask
flask-restful
ramlfications
dill
pandas
pyyaml
